requirements.txt
# Core dependencies for Smart Accessible Routing System
numpy==1.24.3
pandas==2.0.3
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning dependencies
scikit-learn==1.3.0
joblib==1.3.2

# Map data processing
osmnx==1.3.0
geopandas==0.13.2
folium==0.14.0
shapely==2.0.1

# Web framework and API
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4

# Database
sqlite3

# Authentication and user management
flask-login==0.6.2
werkzeug==2.3.6

# Data visualization and mapping
plotly==5.15.0
dash==2.11.1

# Real-time data processing
requests==2.31.0
python-dotenv==1.0.0

# Machine learning for route optimization
scikit-learn==1.3.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0